extract_load/pyproject.toml
[project]
name = "sql-extract-load"
version = "0.1.0"
description = "SQL Extract Load Script"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.61",
    "pandas>=2.3.3",
    "pendulum>=3.1.0",
    "psycopg[binary,pool]>=3.2.12",
    "pyarrow>=22.0.0",
    "sqlalchemy>=2.0.44",
    "boto3-stubs[essential]>=1.40.61"
]

[dependency-groups]
dev = [
    "pyarrow-stubs>=20.0.0.20250928",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.2",
]
